+++ src/main/java/az/code/springweb/services/Random.java
@@ -1,14 +1,14 @@
 package az.code.springweb.services;
 
 public class Random {
-    public static int id = 1;
-    public static int gradeId = 1;
+    public static long id = 1;
+    public static long gradeId = 1;
 
-    public static int getId() {
+    public static long getId() {
         return id++;
     }
 
-    public static int getGradeId() {
+    public static long getGradeId() {
         return gradeId++;
     }
 }
